requirements.txt
apturl==0.5.2
asgiref==3.2.10
bcrypt==3.1.7
beautifulsoup4==4.9.3
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
coloredlogs==14.0
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1.2
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.3.0
django-crispy-forms==1.9.2
django-filter==2.4.0
djangorestframework==3.12.1
duplicity==0.8.12.0
entrypoints==0.3
evdev==1.3.0
fasteners==0.14.1
Flask==1.0.2
future==0.18.2
gunicorn==20.0.4
httplib2==0.14.0
humanfriendly==8.2
idna==2.8
imageio==2.9.0
img2pdf==0.3.6
imutils==0.5.3
itsdangerous==1.1.0
Jinja2==2.11.2
keyring==18.0.1
kiwisolver==1.2.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.5.1
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3
MarkupSafe==1.1.0
matplotlib==3.2.2
monotonic==1.5
MouseInfo==0.1.3
netifaces==0.10.4
networkx==2.4
notify2==0.3
numpy==1.19.0
oauthlib==3.1.0
ocrmypdf==10.2.0
olefile==0.46
opencv-python==4.3.0.36
paramiko==2.6.0
pdfminer.six==20200517
pexpect==4.6.0
pikepdf==1.17.0
Pillow==7.0.0
pluggy==0.13.1
protobuf==3.6.1
PyAutoGUI==0.9.52
pycairo==1.16.2
pycparser==2.20
pycryptodome==3.9.8
pycups==1.9.73
PyGetWindow==0.0.9
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMsgBox==1.0.8
PyNaCl==1.3.0
pynput==1.7.1
pyparsing==2.4.7
PyPDF2==1.26.0
pyperclip==1.8.0
PyQt5==5.14.1
PyRect==0.1.4
pyRFC3339==1.1
PyScreeze==0.1.26
python-apt==2.0.0+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-xlib==0.28
python3-xlib==0.15
PyTweening==1.0.3
pytz==2019.3
PyWavelets==1.1.1
pyxdg==0.26
PyYAML==5.3.1
redis==3.5.3
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
rq==1.5.0
scikit-image==0.17.2
scipy==1.5.1
SecretStorage==2.3.1
setproctitle==1.1.10
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sortedcontainers==2.2.2
soupsieve==2.0.1
sqlparse==0.4.1
systemd-python==234
tesserocr==2.5.1
tifffile==2020.7.4
tqdm==4.47.0
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
Werkzeug==0.15.1
wxPython==4.0.7
xkit==0.0.0
